refactor: implement types/index.ts, Update export type
